zoomToBoundingBox zoom bug fix
There was a problem in some specific phone devices that the calculated zoom give a bigger number than the maximum zoom for the selected Map Tile Provider then zoom was never happening. 

In my case the calculated zoom was giving 19 as value but maximum zoom for my map was 18 so the result was that map was not zooming at all even if the map was already draw and all properly loaded.